Added "No Leagues" display to Listing Page

+++ application/views/admin/league/list.php
@@ -1,6 +1,5 @@
 <?php
-echo $pagination;
-if (count($leagues) > 0) { ?>
+echo $pagination; ?>
     <table>
         <thead>
             <tr>
@@ -11,16 +10,22 @@ if (count($leagues) > 0) { ?>
         </thead>
         <tbody>
     <?php
-    foreach ($leagues as $league) { ?>
+    if (count($leagues) > 0) {
+        foreach ($leagues as $league) { ?>
             <tr>
                 <td><?php echo $league->short_name; ?></td>
                 <td><a href="/admin/league/edit/id/<?php echo $league->id;?>"><?php echo $this->lang->line('league_edit'); ?></a></td>
                 <td><a href="/admin/league/delete/id/<?php echo $league->id;?>"><?php echo $this->lang->line('league_delete'); ?></a></td>
             </tr>
     <?php
+        }
+    } else { ?>
+            <tr>
+                <td colspan="3"><?php echo $this->lang->line('league_no_leagues'); ?></td>
+            </tr>
+    <?php
     } ?>
         </tbody>
     </table>
 <?php
-}
 echo $pagination; ?>
